api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
